requirements.txt
bitstring
numpy
opencv-python
Pillow